Centipedes are updated outside of event handlers. Workaround for QGraphicsEllipse draw issue with unnormalized rectangle

+++ src/Centipede/CentipedeSegmentConstantYREllipse.h
@@ -31,7 +31,7 @@ public:
 signals:
   
   /// Send new geometry for later updating
-  void signalUpdateAngles (double, double);
+  void signalUpdateAngles (int, int);
   
 private:
   CentipedeSegmentConstantYREllipse();
